♻️ Refactor: follow 기능 한 파일에 묶음
